+++ tests/integ/scala/test_function_suite.py
@@ -189,6 +189,7 @@ from snowflake.snowpark.functions import (
     to_geometry,
     to_json,
     to_object,
+    to_time,
     to_timestamp,
     to_timestamp_ltz,
     to_timestamp_ntz,
@@ -1283,6 +1284,70 @@ def test_to_timestamp(session):
             Row(datetime(2020, 4, 5, 2, 3, 4)),
         ],
     )
+
+
+@pytest.mark.localtest
+def test_to_time(session):
+    # basic string expr
+    df = TestData.time_primitives1(session)
+    Utils.check_answer(
+        df.select(*[to_time(column) for column in df.columns]),
+        [
+            Row(time(1, 2, 3)),
+            Row(time(22, 33, 44)),
+        ],
+    )
+
+    # string expr with format
+    df = TestData.time_primitives2(session)
+    Utils.check_answer(
+        df.select(*[to_time(column, "HH12.MI-SS PM") for column in df.columns]),
+        [
+            Row(time(13, 2, 3)),
+            Row(time(22, 33, 44)),
+            Row(time(12, 55, 19)),
+        ],
+    )
+
+    # timestamp
+    df = TestData.time_primitives3(session)
+    Utils.check_answer(
+        df.select(*[to_time(column) for column in df.columns]),
+        [
+            Row(time(12, 13, 14)),
+            Row(time(20, 21, 22)),
+            Row(time(21, 20, 19)),
+        ],
+    )
+
+    # variant type
+    df = TestData.time_primitives4(session)
+    Utils.check_answer(
+        df.select(*[to_time(column) for column in df.columns]),
+        [
+            Row(time(1, 2, 3)),
+            Row(time(21, 20, 19)),
+            Row(
+                None,
+            ),
+        ],
+    )
+
+    # variant data
+
+    # invalid input for string expr with format
+    # TODO: local test error experience SNOW-1235716
+    # currently local testing throws ValueError while live connection throws SQLException
+    # df = session.create_dataframe(
+    #     [("asdfgh",), ("qwerty",)]
+    # ).to_df("a")
+    # Utils.check_answer(
+    #     df.select(to_time("A", "HH12.MI-SS PM")),
+    #     [
+    #         Row(time(13, 2, 3)),
+    #         Row(time(22, 33, 44)),
+    #     ],
+    # )
 
 
 @pytest.mark.localtest