+++ tests/utils.py
@@ -1,5 +1,9 @@
 #!/usr/bin/env python3
 #
+# Copyright (c) 2012-2024 Snowflake Computing Inc. All rights reserved.
+#
+
+#
 # Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
 #
 
@@ -786,6 +790,54 @@ class TestData:
                 StructField("timestamp_ntz", TimestampType(TimestampTimeZone.NTZ)),
                 StructField("timestamp_ltz", TimestampType(TimestampTimeZone.LTZ)),
                 StructField("timestamp_tz", TimestampType(TimestampTimeZone.TZ)),
+            ]
+        )
+        return session.create_dataframe(data, schema)
+
+    @classmethod
+    def time_primitives1(cls, session: "Session") -> DataFrame:
+        # simple string data
+        data = [("01:02:03",), ("22:33:44",)]
+        schema = StructType([StructField("a", StringType())])
+        return session.create_dataframe(data, schema)
+
+    @classmethod
+    def time_primitives2(cls, session: "Session") -> DataFrame:
+        # string data needs format
+        data = [
+            ("01.02-03 PM",),
+            ("10.33-44 PM",),
+            ("12.55-19 PM",),
+        ]
+        schema = StructType([StructField("a", StringType())])
+        return session.create_dataframe(data, schema)
+
+    @classmethod
+    def time_primitives3(cls, session: "Session") -> DataFrame:
+        # timestamp data
+        data = [
+            (datetime(2024, 2, 1, 12, 13, 14),),
+            (datetime(2017, 2, 24, 20, 21, 22),),
+            ("1712265619",),
+        ]
+        schema = StructType(
+            [
+                StructField("a", TimestampType(TimestampTimeZone.NTZ)),
+            ]
+        )
+        return session.create_dataframe(data, schema)
+
+    @classmethod
+    def time_primitives4(cls, session: "Session") -> DataFrame:
+        # variant data
+        data = [
+            ("01:02:03",),
+            ("1712265619",),
+            (None,),
+        ]
+        schema = StructType(
+            [
+                StructField("a", VariantType()),
             ]
         )
         return session.create_dataframe(data, schema)